Refactor common CLI functionality into a new base module so it can be re-used in other StackStorm CLI based apps.
